requirements.txt
appdirs==1.4.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.11.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-filter==1.0.4
django-frontend-template==1.2.0
django-guardian==1.4.9
django-pagedown==0.1.3
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
djangorestframework-jsonapi==2.2.0
djangorestframework-jwt==1.11.0
drf-json-api==0.1.1
futures==3.1.1
idna==2.5
inflection==0.3.1
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.2.1
psycopg2==2.7.2
PyJWT==1.5.2
pyparsing==2.2.0
pytz==2017.2
reportlab==3.4.0
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
tornado==4.5.1
tornado-psycopg2==0.3.1
uritemplate==3.0.0
urllib3==1.21.1
